requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
APScheduler==3.6.3
attrs==24.2.0
beautifulsoup4==4.12.3
cachetools==4.2.2
certifi==2024.8.30
charset-normalizer==3.4.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
setuptools==75.2.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
tornado==6.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
yarl==1.15.5